requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
serpapi
python-dotenv
requests
yt-dlp
playwright
sqlalchemy[asyncio]
aioodbc
passlib[bcrypt]==1.7.4
aioodbc
pydantic[email]
itsdangerous